Move baseurl functionality to url.ts
